backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
ultralytics==8.3.93
Pillow==11.1.0
numpy==2.0.2
matplotlib==3.10.1  # Optionnel, pour le mode debug